+++ sandbox/vue-library/src/components/CedarLintFixture.vue
@@ -235,6 +235,40 @@ export default defineComponent({
     </section>
 
     <section class="fixture-group">
+      <h3>Landing lead</h3>
+      <div class="fixture-stack">
+        <div
+          class="cdr-split-surface cdr-split-surface--bottom cdr-landing-lead"
+        >
+          <div class="cdr-split-surface__top">
+            <img
+              class="cdr-image cdr-landing-lead__image"
+              src="https://cedar.rei.com/rei-passage-2-tent.jpg"
+              alt="A person in a tent"
+            />
+          </div>
+          <div class="cdr-split-surface__bottom">
+            <header
+              class="cdr-heading-subheading-block cdr-landing-lead__copy-block"
+            >
+              <h1
+                class="cdr-heading-display cdr-heading-display--scale-4 cdr-heading-subheading-block__heading"
+              >
+                Camp with no reservations.
+              </h1>
+              <span
+                class="cdr-subheading-sans cdr-heading-subheading-block__subheading"
+              >
+                We have the know-how to equip you to camp in the middle of
+                nowhere.
+              </span>
+            </header>
+          </div>
+        </div>
+      </div>
+    </section>
+
+    <section class="fixture-group">
       <h3>Images</h3>
       <div class="fixture-stack">
         <img